int->qulonglong for partitions sizes 4.2branch
BUG: 133588


svn path=/branches/KDE/4.2/kdeutils/kdf/; revision=919266

+++ kdfwidget.cpp
@@ -81,7 +81,7 @@ int CListViewItem::compare ( Q3ListViewItem *i, int column, bool ) const
   }
 }
 
-void CListViewItem::setKeys (int kb_size, int kb_avail, float percent_full)
+void CListViewItem::setKeys (qulonglong kb_size, qulonglong kb_avail, float percent_full)
 {
   size  = kb_size;
   avail = kb_avail;
